+++ projects/ngx-xchange-icons/src/lib/send.component.ts
@@ -1,0 +1,26 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'xc-icon-send',
+  standalone:true,
+  template: `
+<svg width="46" height="47" viewBox="0 0 46 47" fill="none" xmlns="http://www.w3.org/2000/svg">
+<g clip-path="url(#clip0_2998_16768)">
+<path d="M42.2283 23.4998H21.4865" stroke="#111111" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"/>
+<path d="M42.2282 23.4968L16.7724 35.7533L21.4865 23.4968L16.7724 11.2403L42.2282 23.4968Z" stroke="#111111" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"/>
+</g>
+<defs>
+<clipPath id="clip0_2998_16768">
+<rect width="32" height="32" fill="white" transform="translate(23.3721 0.871094) rotate(45)"/>
+</clipPath>
+</defs>
+</svg>
+
+
+`,
+  styles: [
+  ]
+})
+export class XcIconSendComponent {
+
+}