+++ projects/ngx-xchange-ui/src/directives/input/styles.ts
@@ -1,0 +1,18 @@
+export const APP_TAILWIND_STYLES: { [key: string]: string } = {
+inputPrimary: `
+flex
+items-center
+w-full
+border-2
+border-base
+px-4
+bg-white
+text-Secondary_Fonts
+text-input
+py-2
+font-manrope
+font-regular
+leading-[22.4px]
+rounded-lg
+`
+};